move mappings to container, customizable changes to modal via props
