+++ src/components/CustomModal.js
@@ -1,53 +1,40 @@
 import React from "react";
 
-import Input from "./Input";
 import {
 	withStyles
 } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import Modal from '@material-ui/core/Modal';
 import Button from '@material-ui/core/Button';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import {openModal, closeModal} from '../actions/ModalActions';
+
+
 class CustomModal extends React.Component {
 
 
 
 	constructor(props) {
 		super(props);
-		this.modalStyle = this.getModalStyle();
 		this.state = {
 			open: false
 		}
 	}
 
-	getModalStyle() {
-		const top = 50 + Math.round(Math.random() * 20) - 10;
-		const left = 50 + Math.round(Math.random() * 20) - 10;
-		return {
-			top: `${top}%`,
-			left: `${left}%`,
-			transform: `translate(-${top}%, -${left}%)`,
-		};
-	}
-
 	render() {
 		return (
 			<div>
-      <Button color="primary" onClick={this.props.openModal}>Open Modal</Button>
+      <Button color="primary" onClick={this.props.onClick}>Open Modal</Button>
       <Modal
         aria-labelledby="simple-modal-title"
         aria-describedby="simple-modal-description"
-        open={this.props.open}
-        onClose={this.props.closeModal}
+        open={this.props.isActive}
+        onClose={this.props.onClose}
       >
-        <div style={this.modalStyle} className={this.props.classes.paper}>
+        <div style={this.props.style} className={this.props.classes.paper}>
           <Typography variant="h6" id="modal-title">
-            Text in a modal
+           {this.props.title}
           </Typography>
           <Typography variant="subtitle1" id="simple-modal-description">
-            Duis mollis, est non commodo luctus, nisi erat porttitor ligula.
+           {this.props.body}
           </Typography>
         </div>
       </Modal>
@@ -56,19 +43,7 @@ class CustomModal extends React.Component {
 	}
 }
 
-const mapStateToProps = (state) => {
-	return {
-		open: state.open
-  };
-};
 
-
-const mapDispatchToProps = (dispatch) => {
-	return bindActionCreators({
-		openModal: openModal,
-		closeModal: closeModal
-	}, dispatch);
-}
 
 export default withStyles(theme => ({
 	paper: {
@@ -79,4 +54,4 @@ export default withStyles(theme => ({
 		padding: theme.spacing(4),
 		outline: 'none',
 	},
-}))(connect(mapStateToProps, mapDispatchToProps)(CustomModal));
+}))(CustomModal);
