+++ src/containers/LoginPage.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import Form from '../components/Form';
-import CustomModal from "../components/CustomModal";
 import {
 	Redirect
 } from "react-router-dom";
