+++ src/reducers/ModalReducer.js
@@ -2,11 +2,11 @@ export default function Modal(state, action) {
 
   if (action.type === "OPEN_MODAL") {
     return Object.assign({}, state, {
-      open: true
+      isActive: true
     });
   } else if (action.type === "CLOSE_MODAL" || typeof state === 'undefined') {
     return Object.assign({}, state, {
-      open: false
+      isActive: false
     })
   }
   return state
